feat(desktop): Add base for desktop app
Adding the base code for the desktop application. This Kotlin application
utilizes JavaCV and its Tesseract-OCR libraries to convert an image of source
code into text, then outputs that text to the program output. This program
assumes the first parameter of the application is the file path to the image
of the source code to process.