feat(desktop): Generate basic code architecture for C++ code
The basic class architecture for C++ source code is now generated from the
OCR-converted text that's based on the images acquired by the program. The
classes defined, referenced, and derived from will be tracked through the
parsed source code.

Once all the source code text files are traversed through, a final report will
be logged to the output of the classes found, along with their parent classes
and the other classes referenced within them.