improve(desktop): Output text file organization
The OCR generated text is now saved to files that have the extension of either
".hpp", ".java", or ".kt" for the C++, Java, and Kotlin code parsed
respectively. These output files are also now categorized under the same parent
folder as the original image file, and this parent folder will still reside
under the `output` folder.