feat(desktop): Now saving OCR threshold image
Tesseract-OCR preprocesses the input image before performing recognition on it
to extract the text from the image. This can involve steps like inverting the
image color, converting the image to grayscale, dilating the image, etc.

The preprocessed image produced by Tesseract-OCR is now saved to the same output
folder and filename as its corresponding generated text file, but with the
extension ".png".