feat(desktop): Support screen capturing mode
Screen-captures can now be utilized instead of code file paths for OCR
processing and generating the class hierarchy from the user's code present on
the screen. To trigger this mode, use the CLI "screen" for the 1st positional
parameter instead of the path to the source code files. Once this mode is
triggered, press the "Enter" key on the keyboard to capture and process each
subsequent screen capture desired.

The processed screen capture results will be output to the same text and image
files as the processed results from the previous mode does. The only difference
in this mode is that the base name of the files will be "screen-code" followed
by a numeric counter value starting at 1, and incrementing for each subsequent
screen capture.