fix(desktop): Tesseract image conversion process
Crashes were observed from within Tesseract when providing the library images
that were not of a native resolution. Therefore, the screen area captured is
now padded to a 16:9 resolution if it's not already detected at that resolution.

Also, crashes were observed during OCR processing within Tesseract from the
image data converted from the screen captured image to the format the Tesseract
library expects it in. This happened on subsequent screen captures after the
first capture, which is now fixed as well.