feat(desktop): Support ROI OCR processing
ROI OCR processing is now supported during screen capturing where the 2nd
largest contoured area of the image will be utilized for OCR processing. The
captured image is first turned into a threshold image by utilizing the threshold
value given into the new CLI `<image-threshold-value>`. This value should be
provided from the user in the 3rd positional CLI parameter, and can be a double
value ranging from `0.0` - `255.0`. The largest contours of the image are found
based on the threshold image, where the largest contour is assumed to be the
border of the entire image.

Tesseract is provided the screen-captured image in grayscale with the ROI of the
2nd largest contour. Providing Tesseract the threshold image produced while
finding the contours yields poor OCR results as Tesseract does not do its own
pre-processing on the threshold image beyond color inversion if necessary.
Providing the grayscale image instead allows Tesseract to apply other
pre-processing techniques such as dilation/erosion to produce its own threshold
image that produces more ideal character representations for its OCR processing.

Additionally, the Tesseract generated threshold image that's saved to the output
folder has been renamed to `tesseract-threshold.png`.