feat(app): Implement AR class model rendering functionality
The mobile app view has been rearranged to include an AR view that takes up
70% of the bottom of the screen. This view shows the back camera view, which
it processes for AR spatial and lighting awareness. The previous code image
and text split-view is now at the top 30% of the screen and is in a horizontal
orientation instead of a vertical orientation.

When the user clicks on the desired code image to process the OCR for, the
processed text and a text-version of the class hierarchy will appear in the text
views, while 3D AR models will now be rendered in the AR view. The 3D models
will consist of a transparent red square representing the class that was found
from the processed OCR results with its corresponding class name text floating
above the 3D model. A green 3D circle will be rendered within the class model
if it contains references within the class from previously processed classes.
A smaller blue square will be rendered within the class model if it inherits
from a parent class which has been previously processed.

Additionally, all class models will show 3D line connections to any classes they
either reference or inherit from. These lines will be colored the corresponding
green or blue color based on if it's a reference or parent class connection
respectively. Similarly, the text-form of the class hierarchy has been updated
to color the text of the corresponding class names to match this coloring theme
for what they resemble.

Two buttons have also been implemented in the AR UI view. The first shows the
text "Capture", which is intended to be utilized for capturing images from the
camera and utilizing them to process through the OCR and AR model rendering.
However, this feature isn't fully implemented yet, therefore when it is clicked
it will instead show a Toast text saying just that. However, the other button
implemented initially shows the text "Code | Kotlin". This indicates the current
sample code images that are shown and that can be utilized are the Kotlin code
samples. This can be changed by clicking on this button, which will open a
fly-out menu that shows 3 options in "Kotlin", "Java", and "CPP". When one of
the fly-out menu items are selected, it will update the sample code images
shown and that can be utilized to the corresponding code language associated
with the selected menu item.