fix(app): Kotlin list-view items
When selecting a code language other than `Kotlin` in the UI to display for the
sample images, then selecting the code language `Kotlin`, the Kotlin sample
images were not displaying. Instead a yellow background was displaying with no
items. This has been fixed to correctly show the Kotlin code images again.